+++ Debaser/Attributes/DebaserRevisionAttribute.cs
@@ -3,13 +3,11 @@
 namespace Debaser.Attributes
 {
     /// <summary>
-    /// Include as an update criteria the requirement that this particular property has a value that is incremented compared to the previous value
+    /// Specify that the property (which must be an integer type, either <see cref="short"/>, <see cref="int"/>, or <see cref="long"/>)
+    /// is used to indicate the revision of the row, which is then assumed to be stored in an append-only fashion.
     /// </summary>
-    [AttributeUsage(AttributeTargets.Class, AllowMultiple = true)]
-    public class DebaserRevisionAttribute : DebaserUpdateCriteriaAttribute
+    [AttributeUsage(AttributeTargets.Property)]
+    public class DebaserRevisionAttribute : Attribute
     {
-        public DebaserRevisionAttribute(string propertyName) : base($"[S].[{propertyName}] > [T].[{propertyName}]")
-        {
-        }
     }
 }